requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
astroid==2.12.13 ; python_full_version >= '3.7.2'
black==22.12.0
certifi==2022.12.7 ; python_version >= '3.6'
cfgv==3.3.1 ; python_full_version >= '3.6.1'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
dill==0.3.6 ; python_version >= '3.11'
distlib==0.3.6
django==4.1.4
django-countries==7.5
django-stubs==1.13.1
django-stubs-ext==0.7.0 ; python_version >= '3.7'
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==1.8.0
filelock==3.8.2 ; python_version >= '3.7'
flake8==6.0.0
identify==2.5.11 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
isort==5.11.4 ; python_full_version >= '3.7.0'
lazy-object-proxy==1.8.0 ; python_version >= '3.7'
mccabe==0.7.0 ; python_version >= '3.6'
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
pathspec==0.10.3 ; python_version >= '3.7'
platformdirs==2.6.0 ; python_version >= '3.7'
pre-commit==2.20.0
psycopg2==2.9.5
pycodestyle==2.10.0 ; python_version >= '3.6'
pyflakes==3.0.1 ; python_version >= '3.6'
pyjwt==2.6.0 ; python_version >= '3.7'
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7 ; python_full_version >= '3.6.2'
pytz==2022.7
pyyaml==6.0 ; python_version >= '3.6'
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
setuptools==65.6.3 ; python_version >= '3.7'
sqlparse==0.4.3 ; python_version >= '3.5'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1 ; python_version >= '3.7'
tomlkit==0.11.6 ; python_version >= '3.6'
types-pytz==2022.7.0.0
types-pyyaml==6.0.12.2
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
virtualenv==20.17.1 ; python_version >= '3.6'
wrapt==1.14.1 ; python_version >= '3.11'
